Moved the debug port selector and updated changelog
