+++ .changes/next-release/Feature-cf59acbb-0c3a-4299-9996-c9f83326353e.json
@@ -1,4 +1,4 @@
 {
     "type": "Feature",
-    "description": "Toolkit dynamically chooses an open port when debugging SAM applications"
+    "description": "Toolkit dynamically chooses an available port when debugging SAM applications, starting at port 5858 and couting upwards until one is found"
 }
