+++ src/shared/utilities/vsCodeUtils.ts
@@ -2,7 +2,7 @@
  * Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-import { getPortPromise } from 'portfinder'
+
 import * as vscode from 'vscode'
 import * as nls from 'vscode-nls'
 import { getLogger, Loggable, LogLevel } from '../logger'
@@ -114,9 +114,4 @@ export function getChannelLogger(channel: vscode.OutputChannel): ChannelLogger {
                 templateTokens
             })
     })
-}
-
-export async function getStartPort(): Promise<number> {
-    // should we let the user configure the starting port?
-    return getPortPromise({ port: 5858 })
 }
